Include grouping insights in expand cardinally mine probability calculation
I think a lot of this domain-specific minesweeper strategy could be generalized into geometry/topology problems. If we consider the playing field as a 2D matrix of proximal mine counts, it can be imagined as an elevation map. Then, instead of choosing 'direct groupers' or 'indirect groupers' or 'expansion contenders with grouping possibilities', we can say 'continue eroding the lowest valley'. Personally, I find it much easier to visualize and manipulate with me mind..