Attempt Dos, revision uno: act using a heat map of mine probabilities
Stage 1: we gather all possible sets of unrevealed cells and the number of mines each set contains; distill these until there are no duplicate or sub-groups; for each cell, record the highest mine probability of all the groups it's found in

Stage 2: reveal all cells with a probability of 0, flag all cells with a probability of 1; if no 0's or 1's are found, reveal a random cell; repeat Stage 2 till win/lose

Next up is improving the strategy when no 0's and 1's can be found. Topology is the buzzword of the day.