Rolled some stuff into System... and it looks like all things will get rolled up into Groups, which are just observations from different vantage points. Sometimes only from viewing the problem from many vantage points can we piece together decisive action
Oh, and, uh, started logging the actions Director Dos (hereby known as Director Don AKA Dirty Dan). Also, included some logic for maximum possible mines which can be useful in some endgame cases.

It occurred to me that even if a group isn't a strict subset, it can still offer insight into the board. For instance, if a #1 is in the corner of a bubble, and #2 is its neighbour in the bubble, sharing two of the 1's five unrevealed neighbours, we can still say the one cell #2 has in uncommon with #1 definitely has a cell... because putting two inside the #1's cells would be invalid. More to learn :D